Учтены рекомендации ревьюера: вынесены enum/реализована addSubviews()
