+++ ImageFeed/Resources/Colors.swift
@@ -1,0 +1,14 @@
+//
+//  Colors.swift
+//  ImageFeed
+//
+//  Created by Наталья Черномырдина on 12.04.2025.
+//
+import UIKit
+
+enum Colors {
+    static let black = UIColor(named: "YP Black")
+    static let white = UIColor(named: "YP White")
+    static let gray = UIColor(named: "YP Gray")
+    static let red = UIColor(named: "YP Red")
+}
